news_crawler/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.11.1
blinker==1.8.2
certifi==2022.12.7
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_7a9c7wyorr/croot/cffi_1714483157752/work
charset-normalizer==2.1.1
click==8.1.7
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_51oymkxnaj/croot/cryptography_1724940565328/work
et-xmlfile==1.1.0
fastapi==0.112.0
Flask==3.0.3
h11==0.14.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.2
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
openpyxl==3.0.10
outcome==1.3.0.post0
packaging==24.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL @ file:///Users/cbousseau/work/recipes/ci_py311/pymysql_1678001218899/work
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.28.1
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.37.2
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.13
uvicorn==0.30.5
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.3
wsproto==1.2.0
